Selftests: skip tests if modules and commands are not available
This is an attempt to allow for a more Pythonic development workflow,
allowing Avocado selftests to run on a simpler environment than it's
currently required.

Basically, if one has a system with both Python and setuptools,
(core) Avocado should be able to run.  Conversely, I believe if the
same system has the Avocado source code, after a "python setup.py
develop" of sorts, the selftests should be able to run.

To accomplish this, it's necessary to make the extra requirements (not
defined in the setup.py files) *really* optional while running the
tests.

This removes the requirements to have the "psutil" and "jsonschema"
Python modules, and the "which" binary (usually from the package of
same name).

After this, it's possible to test Avocado from scratch with a workflow
such as:

   $ podman run --rm -ti fedora:35

Then, from inside the container:

   $ dnf -y install python3-setuptools
   $ python3 setup.py develop
   $ python3 setup.py test --skip static-checks

Signed-off-by: Cleber Rosa <crosa@redhat.com>