Test name length: adapt to current filesystem limitations
There has been hardcoded assumptions that all filesystems Avocado
would be operating on would guarantee the 255 limits.  This is an
incorrect assumption.  Filesystems such as overlayfs (used as the
default storage driver in podman) for instance, have smaller limit.

This changes the maximum length of test related files (results
directory and data files) to observe the actual filesystem
limitations.

Signed-off-by: Cleber Rosa <crosa@redhat.com>