Introduce "podman-image" runner and dependency support
Avocado's dependency mechanism exists for the exact purpose of
proactively preparing a test requirement *outside* of the actual test.
This reduces the debugging for test failures which are not within the
scope of what is attempted to be verified, adds time overhead to a
test, etc.

This introduces a "podman-image" runner, which can be used as a
dependency in tests.  This way, the image will be available before the
test runs, or, if it fails to be made available, the test won't be
run.

There's one example in the selftests which embed the requirement of
having the "fedora:34" podman image available.  In the future, I
personally expect that requirements at the job or suite level will
also be supported.  This way, this runner/dependency type would be
able to prepare images before running tests in a podman-based spawner.

Signed-off-by: Cleber Rosa <crosa@redhat.com>