selftests/functional/test_teststmpdir.py: run inner tests serially
The "test_tests_tmp_dir" checks the common temporary directory by
running both a "exec-test" (then known as SIMPLE tests) and an
"avocado-instrumented" (then known as INSTRUMENTED tests).

The first one checks for a single file in the common directory,
while the second checks for exactly two files.

The order in which those checks, and thus the tests, run need to
deterministic.  So, this sets the execution to one task max at
a time.

Signed-off-by: Cleber Rosa <crosa@redhat.com>