lint: address W0221 warnings (arguments-differ)
When the number of method arguments differ on an inherited class,
pylint will give the W0221 warning.

On most cases, these changes are intentional, with base classes
allowing for instance access (bound methods) or the other way around
it.  Exceptionally methods are turned into static versions to avoid
the warning.

Signed-off-by: Cleber Rosa <crosa@redhat.com>