Task timeout handling fix
There is a problem that sometimes the test is interrupted by timeout
before the test even start. Because avocado is creating test metadata
after the test start, and this metadata is necessary for post test
plugins, this behavior can make Avocado crash. This commit will add
metadata creation into the test finish process if the metadata is
missing.

Reference: #5413
Signed-off-by: Jan Richter <jarichte@redhat.com>