exec-test/tap runners: remove instance attribute with runnable
The BaseRunner API was changed so that it can be instantiated once,
and used many times, that is, run() can be called many times with a
runnable instance.

For compatibility sake, most runners still keep a copy of the runnable
in an instance attribute, but that deviates from the design goal of
not keeping state.

This fixes that for the exec-test and tap runners.

Signed-off-by: Cleber Rosa <crosa@redhat.com>