Podman spawner: optmize deployment of Avocado eggs
Given that podman supports mapping volumes to the container, there's
no need to make copies of the eggs into the container image.

This should speed up the podman spawner by saving the two copy
operations.

With this change, there's no need to implement the deploy_avocado()
interface, thus the podman spawner can become a regular Spawner.

The DeploymentSpawner plugin type has not been removed due to the fact
that the capability of mounting files is not universal, and it's
expected that many spawners will need to perform an active deployment
of Avocado.

Signed-off-by: Cleber Rosa <crosa@redhat.com>