Podman spawner: make Python information cache effective
The current use of a dict without any locking means that multiple
instances may attempt to find Python version info, and not find it,
while other(s) instance(s) are already attempting to fetch the info.

With a lock, only the first attempt will cause an actual container
execution and the fetching of the Python information.

A possibility for further optmization is to preserve this information
in the requirement cache database.

Signed-off-by: Cleber Rosa <crosa@redhat.com>