Travis CI: always runs one test at a time
It was observed that some tests on Travis CI simply never return a
result, and jobs get stuck.  This is particularly true for the lint.sh
test, which takes around 200 seconds to finish (when it finishes).
Many times, though, it never reports back to Avocado.

My belief is that Travis may be killing some of the pylint processes,
because of resource exhaustion, and that is never picked up by the
main test script and Avocado.

Let's run one test at a time, so that these situations are minimized
and that it's clear based on the logs which test is not finishing.

Signed-off-by: Cleber Rosa <crosa@redhat.com>