Cache interface
This is an initial commit which crates avocado cache interface. It
introduces new plugin "Cache" which will be responsible for manipulating
with cache, and it's representation through avocado interface.

For now, it supports only listing of cache entries, but it is possible
to expand for features like prune or edit.

Reference: #5453
Signed-off-by: Jan Richter <jarichte@redhat.com>