Work around setuptools bad judgement of package versions
Setuptools is not as smark as pip when it comes to picking a
dependency compatible with the current Python version.  In the
particular case being addressed here, it will attempt to install
the latest markupsafe package, which is not Python 3.6 compatible.

In that case, let's help it by determining the actual version to be
used before it tries install jinja2.

Signed-off-by: Cleber Rosa <crosa@redhat.com>