Drop "nrunner" prefix from config and parameters
The "--nrunner-$option" made sense when the old runner was available.
Nowadays, it's just more typying and confusion to users.

The only argument against keeping the "nrunner" prefix would be if
some options are not implemented by other runners, but still, it can
be considered a limitation of those runners.  In reality, there are no
other runners implemented right now, so that shouldn't be an issue.

Signed-off-by: Cleber Rosa <crosa@redhat.com>