Rename Test Runner to Suite Runner
The original interface name "Runner" comes a time where there was no
concept of multiple test suites per job, and there wasn't any of the
nrunner runners that actually run tests.

The "run_suite()" method makes it clear that this interface actually
is about running suites (with all their tests), but not really a test
or tests directly.

This also renames the "--test-runner" option to "--suite-runner", to
better match its real goal.

Signed-off-by: Cleber Rosa <crosa@redhat.com>