Test runner selection config key and command line option
The "run.test_runner" configuration option will be renamed to
"run.suite_runner" to better match its purpose.  The same is true for
the matching "--test-runner" command line option that will be renamed
to "--suite-runner".

Until then (Avocado 100.0 and later), let's add a deprecation warning
so users are aware of the change.

I attempted the approach of keeping both config keys valid, but
there's no support for aliases in the current settings module.  When
it comes to the command line option, on the other hand, we can have
two options pointing to the same config key, which will allow command
line users a smooth transition.  Config files and job API users will
need to migrate atomically when migrating to version 100.0 or later.

Signed-off-by: Cleber Rosa <crosa@redhat.com>