nrunner specific command line options: drop the nrunner prefix
There's too much verbosity in the nrunner specific command line
options with the "--nrunner-" prefix.  Given that it's the current
default, and it's fair game to believe that it will remain so for some
time, let's simplify user experience and drop that prefix.  Even if
new suite runners are added, they are free to ignore those values.

But, let's also keep the old command line options to avoid disruption.
A deprecation message has been to them, though.

Signed-off-by: Cleber Rosa <crosa@redhat.com>