Python path environment bomb fix
The current setting of PYTHONPATH values is necessary, but as it's
implemented, it is similar to a fork bomb.  On each new execution, new
values are appended to the global PYTHONPATH environment variable.  In
jobs with many tests, this can grow out of the existing
limits (besides adding unnecessary memory overhead).

This was actually caught by failures on COPR builds.

Signed-off-by: Cleber Rosa <crosa@redhat.com>