Runnable: consider configuration used from class definition
Upon a Runnable instantiation, it may need to get the list of
configuration items used if one is not provided.  But, getting that
list by running a command such as "avocado-runner-kind capabilities"
is very expensive.

This adds the definition from the "BaseRunner" attribute, and falls
back to running the command only if such value does not exist.

Reference: https://github.com/avocado-framework/avocado/issues/5495
Signed-off-by: Cleber Rosa <crosa@redhat.com>