nrunner: remove generic "avocado-runner"
The generic "avocado-runner" is a reminescent of the all-in-one
"nrunner.py" file, for ease of deployment.  With the nrunner split and
now deployed via Python eggs, it does not have to exist anymore.

No attempt to deprecate before removal was made because direct usage
of the runners is not expected by most users (they are always executed
by an "avocado run" command).

Signed-off-by: Cleber Rosa <crosa@redhat.com>