Tags: implement filter of resolutions instead of runnable
The filter_test_tags_runnable() function, although named closely after
the version inteded for use with the loader, does not behave the same
way.  It actually only checks one runnable at a time, instead of
actually performing the filtering.

The actual need of the test tag filtering, is to actually work on the
resolutions, and all the runnable within.

This introduces a new utility function, which is now used in the suite
creation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>