Selftests: pin Fedora version used on podman spawner
Because of the release of Fedora 37, the default distro has Python
3.11. The problem with that is there are no "setuptools" (or Avocado
eggs for that matter) available for Python 3.11 yet.

This caused a brekage in the selftests run in CI (or in any
environment where a reference to fedora:latest was refreshed and
resolved to fedora:37).

Because there are tests, it's better to *not* have them changing
behavior because of external events.

This still keeps the default behavior of the spawner to use the latest
fedora if a distro matching the host can not be detected.

Reference: https://github.com/avocado-framework/avocado/issues/5525
Signed-off-by: Cleber Rosa <crosa@redhat.com>