Cache interface only for metadata
When we store images to avocado cache, we store the real data to avocado
file system. On the other hand, when we store test requirements to
avocado cache, we store only the metadata about where and how it was
stored into the file system. The idea for cache interface was to create
unite and common interface for all data in cache. Unfortunately, we
found out that for users, it might be confusing to mix metadata and real
data into one interface. So let's start slowly and provide interface
only for metadata for now.

Reference: #5453
Signed-off-by: Jan Richter <jarichte@redhat.com>