Remove "runner.output.utf8" configuration
This configuration is only used by the varianter plugin deciding how
to "draw" its output.

It should, by default, be based on the current system setting, similar
to what happens with "runner.output.color", which as a default of
"auto" and uses the currently set TERM to determine that.

Unless users require control over this, which I'm not convinced they
do, it should be an automatic decision.

Signed-off-by: Cleber Rosa <crosa@redhat.com>