Job: adjust job start and end time visibility
The current "time_start" and "time_end" attributes of a Job do not
contain useful information, because it holds monotonic times.
This makes those, once public attributes, private ones.

The elapsed time of a job is still accurate, and because of that, is
kept as a public attribute.

Also, a more Pythonic value (None) is used for the attributes which
have not have a value set yet.

Signed-off-by: Cleber Rosa <crosa@redhat.com>