Job: allow for partial automatic calculation of elapsed time
One use case for a job's time accounting is when the start and end
time are set, and then it's expected that the elapsed time will also
be set.  But, right now, the elapsed time is only set if the *end*
time was not previously set.

This changes the behavior, and adds a test that checks that use case.

Signed-off-by: Cleber Rosa <crosa@redhat.com>