selftests/functional/test_job_timeout.py: fix expectations on job timeouts
Depending on the amount of resources on the executing environment, the
runners may not have enough time to even be started and send back
their statuses.

In the "short timeout" test, we can not guarantee that the test
results will be received within the 1 second job timeout, and thus we
can not reliably check against it.

In the "long timeout interrupted" test (a new test), then, it seems to
be possible to expect the tests status, and then we will check against
the presence of the interrupted messages for them.

The previous "long timeout" test (which is now named "long timeout all
ok" which expected the job to complete before the timeout), has been
modified and given more time, so that it can be run more reliably on
every selftest job.

Fixes: https://github.com/avocado-framework/avocado/issues/5544
Signed-off-by: Cleber Rosa <crosa@redhat.com>