Podman spawner: fix outputdir preservation
The Podman spawner creates a volume that serves to preserve a test's
"outputdir" (aka "data") directory.  But, because of possible (and
probable) mismatches between users in the "host" and in the container
itself, the current choice of location for the "outputdir" can prove
ineffective.

This changes the directory used as the outputdir to an absolute
location, that does not depend on the user.

Fixes: https://github.com/avocado-framework/avocado/issues/5556
Signed-off-by: Cleber Rosa <crosa@redhat.com>