Jobs show: remove misleading spawner info
A job is not limited to a single spawner. The spawner is actually used
in the scope of a TestSuite (of which a job can have many). But the
current implementation of "avocado jobs show" shows a single spawner
for a job.

Let's remove that possibly misleading information.  While at it, let's
include an example job that shows multiple spawners in use.

Fixes: https://github.com/avocado-framework/avocado/issues/5553
Signed-off-by: Cleber Rosa <crosa@redhat.com>