exec-test: allow the use of runnable's output dir for stdout/stderr
Most of the information that generic executables (exec-test) tests
generate will be done with their STDOUT and STDERR.

Avocado, since 0a8178ded, inadvertently reduced the size capacity when
keeping and storing the STDOUT and STDERR.  Still, limits will always
exist, either in memory buffer sizes, network buffer sizes, or even
filesystem file size limits or overall storage resources.

This change increases considerably the limits of an "exec-test" by
using, when available, the given "output_dir" parameter to a runnable
or task.  Whenever a test is run under "avocado run", a task will be
used, and a task always has an "output_dir", so, unless one is running
"avocado-runner-exec-test" manually and *not* providing an output_dir,
the STDOUT and STDERR limits are now the filesystem limits.

The tests document the existing limits:

 * 64kib when using PIPEs
 * Considerably larger limits (filesystem ones) when using an
   output_dir

It also "documents" a new behavior introduced here that produces an
error if an output_dir is attempted to be reused, to avoid overwriting
unintended stdout/stderr files.

Fixes: https://github.com/avocado-framework/avocado/issues/5521
Signed-off-by: Cleber Rosa <crosa@redhat.com>