Egg support: be more precise when adding egg location to PYTHONPATH
When running Avocado Python modules, the interpreter on the new
process needs to know where Avocado can be found.  This is usually
handled by the Avocado installation being available to the
standard Python installation, but if Avocado is running from an
uninstalled egg, it needs extra help.

Previously, the content of the set created from the sys.path content
which is not ordered by nature, meant that the order of entries on the
PYTHONPATH would change unpredictably.

Because of the changes in the spawner, and the much smaller intrusion
on PYTHONPATH (and thus on the "import" machinery), the safeloader.sh
test issue goes away.

Fixes: https://github.com/avocado-framework/avocado/issues/5378
Signed-off-by: Cleber Rosa <crosa@redhat.com>