Rename of utils/messages to indicate factory classes
This is simply a name change, but at least for me, it helps a lot
while reading and understanding this code.  The classes being named
ending in "Message" seem to indicate that they are the end result,
that is, actual messages.

My using the Factory prefix, it makes it clear (at least for me) that
these classes will be used to actually manufacture messages.

I've refrained from appending "Factory", that is, having the classes
named as SomethingMessageFactory, because of verbosity given that all
code is container within the "messages" module, and all users of the
code use the "message.SomethingFactory" style.

Signed-off-by: Cleber Rosa <crosa@redhat.com>