Messages: eliminate indirection in the mapping of factories
The dict of supported types of factories (previously message) classes
is not really consulted by any code dynamically.

By referring to the factory classes themselves, instead of
"nicknames", there's no extra hop that one has to follow to have
access to the classes that actually do the work.  IMO this is a
readability improvement, although it's also a reduction in code and
(very minor, but still) performance improvement.

Signed-off-by: Cleber Rosa <crosa@redhat.com>