Podman spawner: use runner plugins instead of removed generic nrunner
On 7a8bbb250, the "avocado-runner" command and the
"avocado.core.nrunner" module entrypoint were removed.  But, the
Podman spawner failed to sync with this change, and kept using the
"avocado.core.nrunner" module when running a task on a container.

The reason this was not caught before is that, all the tests make use
of the previously release egg, and not an egg that matches the
development state.  This should now be protected with a new CI check.

Signed-off-by: Cleber Rosa <crosa@redhat.com>