Pylint: bump to version 2.14.4
Most Avocado developers (that I know, myself included) use a Fedora
workstation for development.  It's usually a good thing to be able to
reuse system packages at development time, if at all possible, and
it's also a good idea to keep refreshing the static code checkers.

This bumps Pylint to 2.14.4, which matches Fedora 37 offerings and
adds new checks.

The reason for the mypy package addition is that, on Pylint 2.14.4,
the "no_self_use" check is an extension, and it requires mypy.

Signed-off-by: Cleber Rosa <crosa@redhat.com>