Provide alternative slot cid or container name through runtime cache
Custom schedulers can still pass a custom spawner handle via the
runtime task but the default avocado state machine will instead be
able to operate without it with the spawner managing the available
and occupied container slots (all of which are also configurable).

Thanks to Cleber for the built-in support for lists in avocado's
core settings, adaptation of the unit tests for avocado runs, as
well as extra mock reset safety for regular python runs.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>
Signed-off-by: Cleber Rosa <crosa@redhat.com>