avocado.utils.archive: identify zstd as archives
This is preliminary work in order to allow for zstd archives to be
compressed and uncompressed by the common API available in
avocado.utils.archive.

It simply adds the ability to check if files seem to be zstd, and adds
that check to the "is_archive" condition.

Another reason for this trivial preliminary work is that it doesn't
seem like even the most basic support for uncompressing zstd archives
will be possible without the use of the zstd tool itself.  This brings
a novelty to this module implementation, in which some level of
graceful degradation or capabilities pattern will be needed.  There is
some prior art on "avocado.utils.iso9660" that can be used on the
complete uncompress/compress implementation.

Reference: https://github.com/avocado-framework/avocado/issues/5609
Signed-off-by: Cleber Rosa <crosa@redhat.com>