avocado.utils.archive: add support for Zstandard uncompression
This adds support for extracting zstd archives using the
avocado.utils.archive.uncompress API, and is similar to the feature
level introduced earlier for lzma archives.

In the future it should be possible to expand this with:

* support for compression
* support for more backends (such as python-zstd or python-zstandard)

All in all, this should be enough to fulfill the use case described
in the matching issue.

Fixes: https://github.com/avocado-framework/avocado/issues/5609
Signed-off-by: Cleber Rosa <crosa@redhat.com>