Selftests: introduce /bin/true and /bin/false substitutes
Some systems do not have a /bin/true and /bin/false binaries.  While
we could use "passtest.py" and "failtest.py", sometimes we want to
test the exec-test bits too.

This introduces a substitute for /bin/true that are used when the
environment is not certain, in contrast with some /bin/true references
that can be found on CI configuration files when the execution is
pinned to a certain system type.

Signed-off-by: Cleber Rosa <crosa@redhat.com>