Possible dependency result
It adds an attribute to RuntimeRask class, which will make dependencies
more configurable. With this, the task could be run, though its
dependencies failed. This will be very useful for post test tasks which
will do the cleanup after tests and they have to be run even then the
test failed.

Signed-off-by: Jan Richter <jarichte@redhat.com>