Selftests: drop "test_" prefix from file names
The "test_" prefix given to the files containing test files is pretty
much a legacy convention from unittest pattern when the test files are
put alongside the modules they test.

In the case of Avocado's selftests file organization, this is not
necessary, and it consumes some "prime real estate" in job reports,
test result directories, etc.

Another reason for doing so is an upcoming proposal that will be sent
to the autils repo with the following organization:

    module: autils/category/module.py
      test: tests/autils/category/module.py
   metadata: metadata/autils/category/module.yaml

When migrating modules from Avocado to autils, the test file name
without the prefix will help.  Mainintaing backports should also be
easier that way.

Signed-off-by: Cleber Rosa <crosa@redhat.com>