Introduce support for test data files in resolvers and spawners
So far, spawners that run tests in different environments have been
mostly guessing about what the test constitutes.

This adds more responsibilities on the resolvers. If applicable, they
can describe which files constitute the test, so that they will be
deployed on the actual environment the test will be executed on.

Signed-off-by: Cleber Rosa <crosa@redhat.com>