Present proper interrupt reason on KeyboardInterrupt/SIGINT
This adds a finer grained handling of the reason why a job is being
interrupted, adding proper support for when the user interrupts it
by pressing CTRL+C (KeyboardInterrupt) or sending a SIGINT signal.

One of the older interrupt tests, currently being skipped, were
re-enabled, with a slight change to reflect to the behavior and
messages given on the UI by nrunner.

Fixes: https://github.com/avocado-framework/avocado/issues/5560
Signed-off-by: Cleber Rosa <crosa@redhat.com>