avocado/utils/ssh.py: add timeout functionality
This uses the remote side "timeout" command to implement a timeout for
the actual command the user intends to run.  The arguments given are
the simplest possible to be made compatible with "timeout"
implementations such as the ones given by busybox.

While it's not possible to have a functional test at this
point (because the test would need a second and properly configured
machine), the example API script should serve as a proof that this
implementation behaves correctly:

  $ python3 examples/apis/utils/ssh_timeout.py
  10.571667929994874
  1.4050129299866967

Signed-off-by: zuohanxu <zuohanxu@uniontech.com>
Signed-off-by: Cleber Rosa <crosa@redhat.com>