External logs
This commit adds support for logs which are not under `avocado.*`
namespace. This will be useful when tests use external libraries with
its own logging namespace. Since avocado doesn't manipulate with root
logger and stores logs only from `avocado.*` namespace, the user has to
specify the external namespaces which he wants to store as well. For
this, he have to use `--store-logging-stream` option.

Reference:
Signed-off-by: Jan Richter <jarichte@redhat.com>