Avocado respects runnable.output_dir
Runnable have output_dir variable, but this is not used when the actual
output dir for a given task is created. This commit changes that.

It will add ability for Runnable to specify the output_dir. It will be
used by some pre- / post-plugins which want to log into the same
directory as their main task. Now they can create runnable with main
task output dir and all logs will be at same place.

Reference: #5647
Signed-off-by: Jan Richter <jarichte@redhat.com>