avocado/core/exceptions.py: drop internal unused exceptions
TestTimeoutInterrupted and TestInterruptedError were once used by the
legacy runner to signal tests that were interrupted either because of
a timeout, or because of a user request (CTRL+C).

The current runner does not depend or uses those exceptions, relying
on status messages to signal those conditions instead.

The TestNotFoundError exception was only used by the old loader, when
it used to present a test that was not found as a test inheriting from
avocado.Test.

Signed-off-by: Cleber Rosa <crosa@redhat.com>