Assets plugin: do not assume assigments have ids
Some "left hand side" part of assigments won't always have an id.
This is the case of a Subscript, that exist in assignments such as:

  foo["bar"] = "baz"

Instead of attempting to deal with each type of "left hand side" here,
let's check for the existence of an "id" attribute, and back off if it
doesn't exist.

Reference: https://docs.python.org/3/library/ast.html#ast.Subscript
Fixes: https://github.com/avocado-framework/avocado/issues/5415
Signed-off-by: Cleber Rosa <crosa@redhat.com>