Address multiprocessing model peculiarities in darwin
Python's multiprocessing module has different methods for starting a new
process.  On darwin/macos, it differs from the default on Linux.

This facilitates the workings of the pickling of a function when running
a process by making it a first level function, and mainly, sets the
"fork" start method on the main "avocado" command, and on all the
"avocado-runner-*" the actually start new processes.

Signed-off-by: Cleber Rosa <crosa@redhat.com>