CI: enable all actual tests on darwin/macos
This puts darwin/macos to the about the same level of coverage (apart
from the specific skips) as the other platforms.

Disclaimer: the core Avocado team has no regular access to those
platforms.  If issues on these platforms arise, either the community
needs to address it, or coverage will be reduced.

Signed-off-by: Cleber Rosa <crosa@redhat.com>