Disable garbage collection during logging
In some circumstances, avocado instrumented runner can hang indefinitely
and block other test to execute. This is caused by a very specific bug
in `multiprocessing.Queue` which can cause deadlock when you're
manipulating with queue during garbage collection. This issue was
investigated and well described by @pevogam in
https://github.com/avocado-framework/avocado/issues/4935#issuecomment-938069202
for long story short, this is the process which causes deadlock:

   1. The program was calling q.put(2).
   2. This involves acquiring a lock.
   3. Half-way through the function call, garbage collection happens.
   4. Garbage collection calls Object.__del__.
   5. Object.__del__ calls q.put(1).
   6. q.put(1) tries to acquire the lock… but the lock is already held,
	so it waits.

Because instrumented runner is using multiprocessing.Queue for storing
logs, this is quite common scenario if any object does logging in
`__del__` method.

This commit disables garbage collection for the process when the logs
are stored to `multiprocessing.Queue`. Such change prevents this kind of
deadlock.

Reference: #4935
Signed-off-by: Jan Richter <jarichte@redhat.com>