Task result in lower case
when we are comparing the task results, we have to be sure that they are
in the same format. Because the format of task result is chosen by the
runner which runs the task, the formats may differ. This change makes
all task results in low-case format. It will fix problem with post test
plugins which might be skipped because of wrong result comparison of
`result` and `satisfiable_deps_execution_statuses`.

Reference: #5681
Signed-off-by: Jan Richter <jarichte@redhat.com>