avocado plugins: gether plugins dynamically instead of manually
The current implementation of "avocado plugins" has a static, hard
coded list of plugins.  This can cause issues with synchronization
between existing plugins and listed plugins, so much that "discoverer"
plugins were not being shown.

This change makes the gathering of plugin information to happen
dynamically, and changes the locality of the plugin type description
to be in the dispatcher itself (a better location IMO).

Signed-off-by: Cleber Rosa <crosa@redhat.com>