Assets plugin: improve speed by using a cache with safeloader results
The assets plugin that will attempt to fetch assets early on a job,
before the tests are actually executed, will:

 * look at a job's test suites
 * look for avocado-instrumented tests
 * save those as possible candidates for having "fetch_asset()" calls
 * individually call fetch_assets() on every one of these candidates

The problem with the last step is that the current implementation uses
a visitor approach that depends on knowing the test methods that exist
in a given test file.  To have that, it will parse the entire file on
every single attempt to find "fetch_asset()" calls on test methods.

The longer a file is (that is, the more avocado.Test classes and
test_* methods it contains), the higher is the waste and the
noticeable performance hit.

Instead of rewriting the plugin (which is really on its way out given
that the dependency functionality is prefered), let's mandate the use
of cache containing the parsing of the test files (done by the
safeloader).

Signed-off-by: Cleber Rosa <crosa@redhat.com>