Reduce IO load from slower aexpect's process pipes
The bulk of runner messages sent to the status server via the network
are also included in the runner's stdout which ends up hogging the
entire aexpect process. Since such stdout is not used anywhere and
our debug.log is already collected via the messages straight to the
status server, dropping it already helps the test perform on par with
equivalent tests from other spawners.

Note that it is not worth implementing additional functionality to
still allow the stdout stream since even if we run `task-run` from
the command line for debugging purposes, it's quite easy to set up
a server and lookt at its output instead. This is even more essential
when it comes to remote / distributed test runs.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>