Task termination fix
Sometimes the tasks are not terminated on the SIGINT call. This is
caused because termination process will leave tasks which are being
handled by starting method.

It happens when termination starts when some task has been removed from
state machine inside start method. In such case, the state machine will
falsely raise `complete` flag and the termination process will leave
this task untouched.

This commit will fix this issue by comparing the number of completed
tasks with the number of all tasks in suite, which will remove the false
`complete` flag.

Signed-off-by: Jan Richter <jarichte@redhat.com>