Avocado-instrumented default timeout removal
This will remove unsystematic timeout inside avocado-instrumented
runner. This timeout has been set to 24 hours and users haven't had a
way how to change it. Because we solved problems with
`task.timeout.running` in #5383 and `job-timeout` in #5295 this default
timeout is not needed anymore. After this change, the users will have
more power over timeouts and will be able to run instrumented tests
longer than 24 hours.

Reference: #5394
Signed-off-by: Jan Richter <jarichte@redhat.com>