Development/build requirements: remove elementpath
In 847d53624, elementpath and xmlschema were added to the build
requirements with the reasoning that coverage was expanded (the tests
that relied on "them" would be enable at build time).  But, in fact,
only xmlschema was being used at the time, and that remains today.

Let's remove elementpath for a cleaner set of build requirements.

Signed-off-by: Cleber Rosa <crosa@redhat.com>