KERNEL: Optimize source code to build the kernel
Signed-off-by: ut003460 <xiongneng@uniontech.com>

Omitted the requirement for multiprocessing. CPU_ Call count() and use None as the default value for njobs.
We will check if njobs is a positive number, and if so, link the '- j' option with its value and add it to
make_ Args list. Then, based on the binary_ Determine whether to add 'deb pkg' to make based on the conditions of
package and distribution name_ Args list. Finally, we use spaces to make_ Connect the strings in the args list and
call the build. make() function to perform the build operation