Whiteboard save removed from Test class
Because of nrunner architecture, it is no need to save whiteboard by
Test class. All logs which have to be stored to files are transmitted by
messages and stored by avocado core. Whiteboard logs had been stored by
Test class and avocado core, which created duplicates in whiteboard
file. Removing it from Test class solves this issue. And we can do it
because saving of whiteboard by Test class is legacy behavior.

Reference: #5770
Signed-off-by: Jan Richter <jarichte@redhat.com>