corrupt resolution result
This commit introduces a new type of resolution result for resolver. It
adds the ability for resolvers to inform the user about files which
might be a test, but there is something wrong inside them.

If the resolver marks the reference as corrupted, the avocado will write
a warning message to the user about it.

Reference: #5710
Signed-off-by: Jan Richter <jarichte@redhat.com>