Spawner: terminate_task returns success or failure
While "terminating a task", Avocado would simply request the
termination, without checking on its success.  This change alters the
interface of the terminate_task Spawner plugin interface when it comes
to the expected return.

Then the statemachine is able to, at the very least, log the
information about the failure to terminate a task.

Signed-off-by: Cleber Rosa <crosa@redhat.com>