Process Spawner: introduce handle class
Some operations to manage the lifecycle of a process running a task
depends on the process itself, and others on the asyncio Task waiting
for the process to finish.

This introduces a class that holds together the process itself and the
asyncio Task that waits for the process to finish in a more convenient
locality.

Signed-off-by: Cleber Rosa <crosa@redhat.com>