Task Termination: introduce a two phase termination
The way the termination of a task happens depends on the isolation
model employed by the spawner.  Still, it's pretty common to have at
least two ways to ask for the termination of running processes,
containers, vms, etc: a nice and a forceful way.

This introduces two configurations that allow the user to tune how
long the spawner should wait between the two.  The first interval,
"from_soft_to_hard_termination", can give tests running on tasks the
chance to clean up.  For tests that implement a tearDown() like
pattern, this can be the time given to do run it.

The second interval is the amount of time one expects the process,
container, vm, etc, to have completed its termination, so that a query
about its status will show it.

Signed-off-by: Cleber Rosa <crosa@redhat.com>