INTERRUPTED status overwrite fix
When avocado-instrumented test is interrupted, the avocado runner will
call a tearDown method to clean the test environment. If some error is
raised during the tearDown phase, the `INTERRUPTED` status is
overwritten by this error. This is not desired behavior, because the
reason for the error was interruption of the test. This fix keeps
logging of the tearDown error without change of the `INTERRUPTED`
status.

Reference: #5801
Signed-off-by: Jan Richter <jarichte@redhat.com>