avocado: Introduce "avocado-classless" plugin
The passt project wants to use avocado for tests, but we don't like
its default jUnit style of test presentation.  So, we plan to use some
custom plugins to make test writing more approachable.  Start with the
implementation of the "runner" plugin.  We also add some supporting
pieces: an interim trivial resolver plugin and setup.py so we can
actually do some stuff with that runner.

As for the resolver for the avocado-classless plugin, it requires you
to give a full function name for every test you want to run.
Implement a more "normal" resolver which will run all suitable tests
matching a regular expression.

In the test files, tests are marked by using a decorator to put them
in a special "manifest" of tests in the file.

With avocado-classless we obviously can't use the built in assertion
methods from unittest / avocado-instrumented.  This includes our own
non-OOP helpers for this.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Cleber Rosa <crosa@redhat.com>