avocado/core/settings.py: respect previously set metavar
When registering an option and setting a metavar, and then later
adding a command line option, the metavar will get reset (to None) if
one is not given again.

Let's respect the previously set metavar, if one exists.

As an example where this manifests, the "--show" option of the
"avocado" command line option, currently lacks a metavar even with one
being set for the setting itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>