Human UI Plugin: increase level priority interrupted message
When a job gets interrupted, users want to notice it.  So far, it's
been presented in the Human UI as a simple "INFO" message, that may
"fade into the background" with the other messages.

By bumping it a level to a warning, it should be easier for users to
notice it.

This is not made into a error, because the caused for the interruption
may be an expected situation, such as a job timeout.

Signed-off-by: Cleber Rosa <crosa@redhat.com>