Drop Python 3.7
Given that, after the release of 103.0 LTS, we're now on track for
bigger chances, let's start by dropping Python 3.7 support.

There are a number of features that are available on Python 3.8 and
later only. One example is the availibility of importlib.metadata
stardand library, which we intend to use to replace pkg_resources (and
then setuptools as a whole).

As for Avocado users and important platforms, we're on a good position
such as QEMU already requiring Python 3.8, and CentOS Stream 9
shipping Python 3.9.18.

For RPM builds, EL8 (in all its forms) are not going to be built
anymore on CI.  This may change if the SPEC file is updated to support
a conditional build using Python 3.8 for EL8, but that depends on the
existence of all dependencies on that build, including
python3-aexpect, which is not currently available.

Reference: https://github.com/avocado-framework/avocado/issues/5754
Signed-off-by: Cleber Rosa <crosa@redhat.com>